TRTWrapper moved to networks
