+++ monai/networks/__init__.py
@@ -12,6 +12,7 @@
 from __future__ import annotations
 
 from .utils import (
+    add_casts_around_norms,
     convert_to_onnx,
     convert_to_torchscript,
     convert_to_trt,
