+++ monai/utils/__init__.py
@@ -152,5 +152,3 @@ from .type_conversion import (
     get_numpy_dtype_from_string,
     get_torch_dtype_from_string,
 )
-
-TRTWrapper, TRT_AVAILABLE = optional_import("monai.utils.trt_utils", name="TRTWrapper")
